Moved types to Geometry namespaces
